server/requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.10
djangorestframework==3.12.4
pymongo==3.12.0
python-dotenv==0.19.0
requests==2.26.0
sqlparse==0.4.2
uritemplate==4.1.1
google-api-python-client==2.33.0
uvicorn==0.13.4
httpx==0.18.1
asyncio==3.4.3
celery==5.2.1
redis==4.0.2